if not os.path.exists(part_type):

+++ python/techvision.py
@@ -224,20 +224,23 @@ class TechVision(threading.Thread):
                             if camera_db.outTrainModeOn[0]:
                                 if camera_db.outPartPresentInNest[0]:
                                     if self.capture is not None:
-                                        nest = cv2.imread(os.path.join(part_type, "nest_" + part_pos_num + ".png"))
-                                        nest_part = self.capture.depth.read()
-                                        if nest_part is not None:
-                                            res, nest_mask = self.subtract_background(nest_part, nest)
-                                        if res:
-                                            res = cv2.imwrite(os.path.join(part_type, part_pos_num + ".png"),
-                                                              nest_mask)
-                                            os.remove(os.path.join(part_type, "nest_" + part_pos_num + ".png"))
-                                        else:
-                                            print('Ошибка. Невозможно обучить деталь')
+                                        if os.path.exists(part_type):
+                                            nest = cv2.imread(os.path.join(part_type, "nest_" + part_pos_num + ".png"))
+                                            nest_part = self.capture.depth.read()
+                                            if nest_part is not None:
+                                                res, nest_mask = self.subtract_background(nest_part, nest)
+                                            if res:
+                                                res = cv2.imwrite(os.path.join(part_type, part_pos_num + ".png"),
+                                                                  nest_mask)
+                                                os.remove(os.path.join(part_type, "nest_" + part_pos_num + ".png"))
+                                            else:
+                                                print('Ошибка. Невозможно обучить деталь')
                                 else:
                                     if self.capture is not None:
                                         part = self.capture.depth.read()
                                         if part is not None:
+                                            if not os.path.exists(part_type):
+                                                os.makedirs(part_type)
                                             res = cv2.imwrite(os.path.join(part_type, "nest_" + part_pos_num + ".png"),
                                                               part)
                                     else:
